pyproject.toml
[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
(
  /(
    | venv
    | _build
    | buck-out
    | build
    | migrations
    | config/settings
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120

[tool.poetry]
name = "club_rusjet"
version = "0.1.0"
description = ""
authors = ["Konstantin Raikhert <raikhert13@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.12"
Django = "^4.0.1"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
Pillow = "^9.0.0"
gunicorn = "^20.1.0"
django-cleanup = "^6.0.0"
django-ckeditor = "^6.4.1"
django-webp-doubles = "^0.3"
django-bootstrap-modal-forms = "^2.2.0"
django-widget-tweaks = "^1.4.12"
pip = "^22.2.2"
install = "^1.3.5"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^6.3.0"}

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
flake8 = "^4.0.1"
black = {version = "^21.12b0", allow-prereleases = true}
django-debug-toolbar = "^3.2.4"
isort = "^5.10.1"
django-extensions = "^3.1.5"
ipython = "^8.0.1"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
